refactor: add scss modules
