+++ src/Pages/Auth/Auth.tsx
@@ -1,0 +1,5 @@
+import styles from './Auth.module.scss';
+
+export const Auth = () => {
+  return <section className={styles.auth}>1</section>;
+};
