+++ src/Pages/Auth/index.tsx
@@ -1,5 +1,1 @@
-import './index.scss';
-
-export const Auth = () => {
-  return <section className="auth">1</section>;
-};
+export * from './Auth';
