+++ src/Pages/GraphQl/GraphQL.tsx
@@ -1,0 +1,5 @@
+import styles from './GraphQL.module.scss';
+
+export const GraphQL = () => {
+  return <section className={styles.graphqL}>3</section>;
+};
