+++ src/Pages/GraphQl/index.tsx
@@ -1,5 +1,1 @@
-import './index.scss';
-
-export const GraphQL = () => {
-  return <section className="graphqL">3</section>;
-};
+export * from './GraphQL';
