+++ src/Pages/NotFound/NotFound.module.scss
@@ -1,3 +1,3 @@
-.not-found {
+.notFound {
   color: orange;
 }
