+++ src/Pages/NotFound/NotFound.tsx
@@ -1,0 +1,5 @@
+import styles from './NotFound.module.scss';
+
+export const NotFound = () => {
+  return <section className={styles.notFound}>404</section>;
+};
