+++ src/Pages/NotFound/index.tsx
@@ -1,5 +1,1 @@
-import './index.scss';
-
-export const NotFound = () => {
-  return <section className="not-found">404</section>;
-};
+export * from './NotFound';
