+++ src/Pages/Welcome/Welcome.tsx
@@ -1,0 +1,5 @@
+import styles from './Welcome.module.scss';
+
+export const Welcome = () => {
+  return <section className={styles.welcome}>2</section>;
+};
