+++ src/Pages/Welcome/index.tsx
@@ -1,5 +1,1 @@
-import './index.scss';
-
-export const Welcome = () => {
-  return <section className="welcome">2</section>;
-};
+export * from './Welcome';
