+++ src/components/footer/Footer.tsx
@@ -1,0 +1,5 @@
+import styles from './Footer.module.scss';
+
+export const Footer = () => {
+  return <footer className={styles.footer}> 000</footer>;
+};
