+++ src/components/footer/index.tsx
@@ -1,5 +1,1 @@
-import './index.scss';
-
-export const Footer = () => {
-  return <footer className="footer">000</footer>;
-};
+export * from './Footer';
