+++ src/main.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import { ThemeProvider } from './providers/ThemeProvider';
-import { App } from './components/app';
+import { App } from './components/App/App';
 import './shared/config/i18n/i18n';
 import './styles/index.scss';
